disp: Switch to using show-image
Keyboard event handling is restrictive in minifb when trying to share
the state outside of the Display module. So, switch to the show-image
library to draw the display.

Update one of the CPU unit tests for display clear, since we really
shouldn't need the Display to be present to test that code path out.