audio: Implement audio module
This uses SDL2 to output a system sound. Make an object using the struct
from this module as a member of the timer module. It is far more
convenient to embed it inside the timer object, than to try to keep it
separate and then pass a reference to the separate object to the timer
object.